add languageTextField visualization
